Move AuthorizationHandler to authorization module
Moves the `AuthorizationHandler` trait and its
`AuthorizationHandlerResult` enum into the authorization module.
