+++ libsplinter/src/rest_api/auth/authorization/maintenance/mod.rs
@@ -22,7 +22,9 @@ use std::sync::{
 };
 
 use crate::error::InternalError;
-use crate::rest_api::auth::{identity::Identity, AuthorizationHandler, AuthorizationHandlerResult};
+use crate::rest_api::auth::identity::Identity;
+
+use super::{AuthorizationHandler, AuthorizationHandlerResult};
 
 /// An authorization handler that allows write permissions to be temporarily revoked
 ///
