+++ libsplinter/src/rest_api/auth/mod.rs
@@ -22,15 +22,13 @@ pub mod identity;
 
 use std::str::FromStr;
 
-#[cfg(feature = "authorization")]
-use crate::error::InternalError;
 use crate::error::InvalidArgumentError;
 
 #[cfg(feature = "authorization")]
 use super::Method;
 
 #[cfg(feature = "authorization")]
-use authorization::PermissionMap;
+use authorization::{AuthorizationHandler, AuthorizationHandlerResult, PermissionMap};
 use identity::{Identity, IdentityProvider};
 
 /// A permission assigned to an endpoint
@@ -46,40 +44,6 @@ pub enum Permission {
     AllowAuthenticated,
     /// Allow any request without checking for authorization.
     AllowUnauthenticated,
-}
-
-/// An authorization handler's decision about whether to allow, deny, or pass on the request
-#[cfg(feature = "authorization")]
-pub enum AuthorizationHandlerResult {
-    /// The authorization handler has granted the requested permission
-    Allow,
-    /// The authorization handler has denied the requested permission
-    Deny,
-    /// The authorization handler is not able to determine if the requested permission should be
-    /// granted or denied
-    Continue,
-}
-
-/// Determines if a client has some permissions
-#[cfg(feature = "authorization")]
-pub trait AuthorizationHandler: Send + Sync {
-    /// Determines if the given identity has the requested permission
-    fn has_permission(
-        &self,
-        identity: &Identity,
-        permission_id: &str,
-    ) -> Result<AuthorizationHandlerResult, InternalError>;
-
-    /// Clone implementation for `AuthorizationHandler`. The implementation of the `Clone` trait for
-    /// `Box<dyn AuthorizationHandler>` calls this method.
-    fn clone_box(&self) -> Box<dyn AuthorizationHandler>;
-}
-
-#[cfg(feature = "authorization")]
-impl Clone for Box<dyn AuthorizationHandler> {
-    fn clone(&self) -> Box<dyn AuthorizationHandler> {
-        self.clone_box()
-    }
 }
 
 /// The possible outcomes of attempting to authorize a client
