+++ services/scabbard/libscabbard/src/client/mod.rs
@@ -503,7 +503,10 @@ mod tests {
     use actix_web::HttpResponse;
     use futures::future::IntoFuture;
     #[cfg(feature = "authorization")]
-    use splinter::rest_api::auth::{AuthorizationHandler, AuthorizationHandlerResult, Permission};
+    use splinter::rest_api::auth::{
+        authorization::{AuthorizationHandler, AuthorizationHandlerResult},
+        Permission,
+    };
     use splinter::{
         error::InternalError,
         rest_api::{
